requirements.txt
flask>=3.0.0
groq>=0.9.0
chromadb>=0.5.0
rank-bm25>=0.2.2
openai>=1.0.0
gitpython>=3.1.0